langpipe/requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
filelock==3.19.1
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.76
langgraph==0.6.8
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.31
lz4==4.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pillow==11.3.0
psutil==7.1.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
safetensors==0.6.2
sentencepiece==0.2.1
sentry-sdk==2.39.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.2
trino==0.336.0
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
xxhash==3.5.0
zstandard==0.25.0
